+++ projects/packages/publicize/_inc/components/customize-and-preview/index.tsx
@@ -1,4 +1,5 @@
 import { useBreakpoint } from '@automattic/viewport-react';
+import { useSyncMediaToConnections } from '../../hooks/use-sync-media-to-connections';
 import { hasSocialPaidFeatures } from '../../utils';
 import { CustomizationToggle } from './customization-toggle';
 import styles from './styles.module.scss';
@@ -13,6 +14,9 @@ import { TabPanelMobile } from './tab-panels/mobile';
 export function CustomizeAndPreview() {
 	const isSmallScreen = useBreakpoint( '<782px' );
 	const hasPaidFeatures = hasSocialPaidFeatures();
+
+	// Sync media URLs (SIG, featured image) to connections when they change
+	useSyncMediaToConnections();
 
 	return (
 		<div className={ styles[ 'customize-and-preview' ] }>